refactor(middleend): split validate logic to a new file
